+++ fe2o3-amqp-management/src/operations/node/get_mgmt_nodes.rs
@@ -33,7 +33,7 @@ impl<'a> GetMgmtNodesRequest<'a> {
     }
 }
 
-impl<'a> Request for GetMgmtNodesRequest<'a> {
+impl Request for GetMgmtNodesRequest<'_> {
     const OPERATION: &'static str = GET_MGMT_NODES;
 
     type Response = GetMgmtNodesResponse;